refactor: delete empty data container component
