+++ app/ui/lecture/lecture-search/index.tsx
@@ -2,14 +2,29 @@
 import { useAtomValue } from 'jotai';
 import LectureSearchBar from './lecture-search-bar';
 import { searchWordAtom } from '@/app/store/search-word';
-import EmptyDataContainer from './empty-data-container';
 import { Suspense } from 'react';
 import { LectureSearchResultSpinner } from './lecture-search-result/lecture-search-result-spinner';
 import LectureSearchResult from './lecture-search-result';
+import Image from 'next/image';
+import searchResultIcon from '@/public/assets/searchResultIcon.svg';
+import List from '../../view/molecule/list';
+
+const emptyDataRender = () => {
+  return (
+    <div className="flex flex-col items-center justify-center gap-4">
+      <Image src={searchResultIcon} alt="search-result-icon" width={40} height={40} />
+      <div className="text-md font-medium text-gray-400 text-center whitespace-pre-wrap">
+        검색 결과가 표시됩니다
+        <br />한 글자 이상 검색해주세요
+      </div>
+    </div>
+  );
+};
 
 export default function LectureSearch() {
   const searchWord = useAtomValue(searchWordAtom);
   const searchable = searchWord.keyword && searchWord.keyword.length > 1;
+
   return (
     <div className="bg-white  w-full h-[500px] sm:h-[400px] z-[10] flex justify-center" data-testid="lecture-search">
       <div className="w-[800px] mx-auto my-7  flex flex-col gap-10 sm:gap-6">
@@ -19,7 +34,7 @@ export default function LectureSearch() {
             <LectureSearchResult />
           </Suspense>
         ) : (
-          <EmptyDataContainer />
+          <List data={[]} isScrollList={true} emptyDataRender={emptyDataRender} />
         )}
       </div>
     </div>
